requirements.txt
# Core Web / API
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic[email]==2.6.3
python-dotenv==1.0.1
httpx==0.27.0

# Machine Learning & Data Science
scikit-learn==1.4.1.post1
xgboost==2.0.3
pandas==2.2.1
numpy==1.26.4
networkx==3.2.1
pyarrow==15.0.1
shap==0.44.1
matplotlib==3.8.3
seaborn==0.13.2

# Security & OSINT
python-whois==0.8.0
dnspython==2.6.1
requests==2.31.0
beautifulsoup4==4.12.3
pyOpenSSL==24.0.0

# Database & Caching
redis==5.0.1
psycopg2-binary==2.9.9

# Application Utils
PyYAML==6.0.1
loguru==0.7.2
tqdm==4.66.2

# Dashboard
streamlit==1.31.1

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
coverage==7.4.3
